--- src/helpers/index.js
@@ -1,5 +1,0 @@
-export const env = (name, value) => (
-  process.env[name] ? process.env[name] : value
-);
-
-export default {};
